Add files via upload
Starting point

This is some already modified Udacity code.  It currently does not work
and the next steps will be figuring out why.

The end goal for this project is to have a Java program that plays 
Connect4 either against the user with either two human players or 
vs an AI opponent.  For now just simple AI opponent will do 
(random moves), but a future extension might entail using machine
learning for the AI opponent.